@@ -6,6 +6,7 @@ import 'dart:developer';
 
 import 'package:flutter/material.dart';
 
+import 'owner.dart';
 import 'owner_details.dart';
 import 'owner_provider.dart';
 
@@ -13,9 +14,9 @@ import 'owner_provider.dart';
  *
  */
 class OwnersPage extends StatefulWidget {
-  OwnersPage({Key key, this.title}) : super(key: key);
+  OwnersPage({Key? key, this.title}) : super(key: key);
 
-  final String title;
+  final String? title;
 
   @override
   _OwnersPageState createState() => _OwnersPageState();
@@ -28,7 +29,7 @@ class _OwnersPageState extends State<OwnersPage> {
   /**
    *
    */
-  OwnerProvider provider;
+  late OwnerProvider provider;
 
   /**
    *
@@ -48,35 +49,39 @@ class _OwnersPageState extends State<OwnersPage> {
     return Scaffold(
       body: FutureBuilder(
         future: provider.getOwners(),
-        builder: (context, snapshot) {
+        builder: (context, AsyncSnapshot snapshot) {
           log("$snapshot.data");
-          final owners = snapshot.data;
+          final owners = snapshot.data ;
           if (snapshot.connectionState == ConnectionState.done) {
-            return ListView.separated(
-              separatorBuilder: (context, index) {
-                return Divider(
-                  height: 2,
-                  color: Colors.black,
-                );
-              },
-              itemCount: owners.length,
-              itemBuilder: (context, index) {
-                return ListTile(
-                  title: Text(
-                      '${owners[index].firstName} ${owners[index].lastName} (${owners[index].telephone})'),
-                  subtitle:
-                  Text('${owners[index].address}, ${owners[index].city} '),
-                  onTap: () {
-                    Navigator.of(context).push(
-                      MaterialPageRoute<void>(builder: (BuildContext context) {
-                        return OwnerDetails(owner: owners[index]);
-                      }),
-                    );
-                    setState(() {});
-                  },
-                );
-              },
-            );
+            if (snapshot.hasData) {
+              return ListView.separated(
+                separatorBuilder: (context, index) {
+                  return Divider(
+                    height: 2,
+                    color: Colors.black,
+                  );
+                },
+                itemCount: snapshot.data!.length, //owners.length,
+                itemBuilder: (context, index) {
+                  return ListTile(
+                    title: Text(
+                        '${owners[index].firstName} ${owners[index]
+                            .lastName} (${owners[index].telephone})'),
+                    subtitle:
+                    Text('${owners[index].address}, ${owners[index].city} '),
+                    onTap: () {
+                      Navigator.of(context).push(
+                        MaterialPageRoute<void>(builder: (
+                            BuildContext context) {
+                          return OwnerDetails(owner: owners[index]);
+                        }),
+                      );
+                      setState(() {});
+                    },
+                  );
+                },
+              );
+            }
           }
           return Center(
             child: CircularProgressIndicator(),
