@@ -6,6 +6,7 @@ import 'dart:developer';
 
 import 'package:flutter/material.dart';
 
+import 'veterinarian.dart';
 import 'veterinarian_provider.dart';
 
 /**
@@ -15,12 +16,12 @@ class VeterinariansPage extends StatefulWidget {
   /**
    *
    */
-  VeterinariansPage({Key key, this.title}) : super(key: key);
+  VeterinariansPage({Key? key, this.title}) : super(key: key);
 
   /**
    *
    */
-  final String title;
+  final String? title;
 
   /**
    *
@@ -36,7 +37,7 @@ class _VeterinariansPageState extends State<VeterinariansPage> {
   /**
    *
    */
-  VeterinarianProvider provider;
+  VeterinarianProvider? provider;
 
   /**
    *
@@ -55,10 +56,10 @@ class _VeterinariansPageState extends State<VeterinariansPage> {
   Widget build(BuildContext context) {
     return Scaffold(
       body: FutureBuilder(
-        future: provider.getVeterinarians(),
-        builder: (context, snapshot) {
+        future: provider?.getVeterinarians(),
+        builder: (context, AsyncSnapshot snapshot) {
           log("$snapshot.data");
-          final veterinarians = snapshot.data;
+          final veterinarians = snapshot.data ;
           if (snapshot.connectionState == ConnectionState.done) {
             return ListView.separated(
               separatorBuilder: (context, index) {
